# student-service/openapi/student-service-openapi.yaml
openapi: 3.0.3
info:
  title: Student Service API
  version: 1.0.0
  description: API for managing student records
servers:
  - url: http://localhost:8081/api
    description: Student Service API
paths:
  /students:
    get:
      summary: Retrieve all students
      operationId: getAllStudents
      responses:
        '200':
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
    post:
      summary: Create a new student
      operationId: createStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '201':
          description: Student created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Invalid request
  /students/{id}:
    get:
      summary: Retrieve a student by ID
      operationId: getStudentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Student details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Student not found
    put:
      summary: Update a student
      operationId: updateStudent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '200':
          description: Student updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Student not found
        '400':
          description: Invalid request
    delete:
      summary: Delete a student
      operationId: deleteStudent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Student deleted
        '404':
          description: Student not found
components:
  schemas:
    Student:
      type: object
      properties:
        studentId:
          type: integer
          format: int64
        studentName:
          type: string
        grade:
          type: string
        mobileNumber:
          type: string
        schoolName:
          type: string
      required:
        - studentId
        - studentName
        - grade
        - mobileNumber
        - schoolName